services/ml-sidecar/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
redis==5.2.1
psycopg2-binary==2.9.10
sentence-transformers==3.4.1
numpy==2.2.2
